refactor(validator): update import paths and remove obsolete types
* Adjusted import paths in various files to reflect the new structure, replacing `main/services/database` with `main/database`.
* Updated the `go.mod` file to correct paths for `main/types` and `main/database`.
* Deleted unused type definitions and validation files related to GTFS rules and caching to streamline the codebase.

+++ modules/plans/apps/validator/validator/validations/shapes/tests/shape_points_coordinates_consistent_validation_test.go
@@ -15,14 +15,14 @@ func TestShapePointsCoordinatesConsistentValidationUsesConfiguredTolerance(t *te
 		{
 			ShapeId:         lib.Ptr("shape-a"),
 			ShapePtSequence: lib.Ptr(1),
-			ShapePtLat:      lib.Ptr(float32(38.0)),
-			ShapePtLon:      lib.Ptr(float32(-9.0)),
+			ShapePtLat:      lib.Ptr(38.0),
+			ShapePtLon:      lib.Ptr(-9.0),
 		},
 		{
 			ShapeId:         lib.Ptr("shape-a"),
 			ShapePtSequence: lib.Ptr(2),
-			ShapePtLat:      lib.Ptr(float32(38.011)),
-			ShapePtLon:      lib.Ptr(float32(-9.0)),
+			ShapePtLat:      lib.Ptr(38.011),
+			ShapePtLon:      lib.Ptr(-9.0),
 		},
 	}
